% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_corr.R
\name{comp.corr}
\alias{comp.corr}
\title{Computing corr}
\usage{
comp.corr(data, timepoints, C)
}
\arguments{
\item{data}{a matrix of data with the columns representing different timepoints}

\item{timepoints}{a vector which contains all the timepoints represented in the data}

\item{C}{a numeric value of C used in computing weighted correlation}
}
\value{
a simmilarity matrix with values between -1 and 1 (1 highly correlated, 0 no correlation )
}
\description{
This function computes the correlation upon picking the best lag and adjusting the data for lags
}
\author{
Thevaa Chandereng, Anthony Gitter
}
