% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_lag.R
\name{weight.lag}
\alias{weight.lag}
\title{Mean Square Distance}
\usage{
weight.lag(x1, x2)
}
\arguments{
\item{x1}{a vector}

\item{x2}{a vector with same length of x1}
}
\value{
a numerical value of difference in squared mean of vector x1 and x2
}
\description{
This function computes difference in mean square distance between two vector
}
\author{
Thevaa Chandereng, Anthony Gitter
}
