% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep.data}
\alias{prep.data}
\title{Preparing Data}
\usage{
prep.data(data, lags)
}
\arguments{
\item{data}{a matrix with columns representing different timepoints}

\item{lags}{a vector of same length as the number of rows in the data column indicating the best lags}
}
\value{
a matrix with NA's introduced where lags are needed
}
\description{
This function prepares the data to compute correlation by introducing NA's when lags are needed
}
\author{
Thevaa Chandereng, Anthony Gitter
}
