% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Example datasets for LPC}
\format{data matrix}
\usage{
simdata
}
\description{
Data is simulated from ImpulseDE function with 8 timepoints c(0, 2, 4, 6, 8, 18, 24, 32, 48, 72)
}
\examples{
data(simdata)
}
\keyword{datasets}
