% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_lag.R
\name{weight.lag}
\alias{weight.lag}
\title{Weight Lag}
\usage{
weight.lag(x1, x2)
}
\arguments{
\item{x1}{a vector}

\item{x2}{a vector}
}
\value{
a matrix with two rows with the shortest length of the vector as the number of columns
}
\description{
This function matches vector of two different lengths
}
\examples{
weight.lag(1:5, 2:9)


}
\author{
Thevaa Chandereng, Anthony Gitter
}
