% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight}
\alias{weight}
\title{Weight in correlation}
\usage{
weight(t, lag, C)
}
\arguments{
\item{t}{a vector of timepoints}

\item{lag}{a integer value of lag used}

\item{C}{a numeric of the constant used in penalty and weight of correlation composed of gaussian kernel}
}
\value{
a list containing w0 and vector w used in computing weighted correlation
}
\description{
This function computes the weight used for correlation using timepoints used and lags used
}
\examples{
weight(t = c(0, 5, 10, 15, 20), lag = 1, C = 20)


}
\author{
Thevaa Chandereng, Anthony Gitter
}
